requirements.txt
mcp[cli]
httpx
python-dotenv
starlette
uvicorn
